Hopefully the program is finished
Changed the function types for small_straight and large_straight from bool to integer. Likewise, the return types true were changed to 1 and the return types false were changed to zero. For some reason, the program works now. I think we can conclude that the gnu g++ compile is greater and all the other g++ compilers are terrible.